% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\docType{data}
\name{players}
\alias{players}
\title{NBA & ABA player directory.}
\format{A data frame with 4529 rows and 10 fields:
\describe{
\item{Player}{Player name}
\item{PlayerId}{Player page Id}
\item{From}{First year in league}
\item{To}{Last year in league}
\item{Position}{Position(s) played}
\item{Height}{Height in 'feet-inches' format}
\item{Weight}{Weight in pounds}
\item{DOB}{Date of birth}
\item{College}{What college did the player attend?}
\item{HOF}{Is this player in the Hall of Fame?}
}}
\source{
\url{https://www.basketball-reference.com/players/}
}
\usage{
players
}
\description{
A dataset containing the
}
\keyword{datasets}
