% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_game_logs.R
\name{scrape_game_logs}
\alias{scrape_game_logs}
\title{Scrape player per game statistics for a given season.}
\usage{
scrape_game_logs(player_id, year)
}
\arguments{
\item{player_id}{The basketball-reference player page identifier code.}

\item{year}{The year that the season in question finished. i.e. 2015 would
return the statistics for the season 2014-15.}
}
\value{
A tibble with a row for each player.

The fields are:
\describe{
\item{Season}{The regular season the game took place in}
\item{Player}{Player name}
\item{G}{Season game}
\item{Date}{The date the game took place on}
\item{Age}{The age of the player on the day of the game}
\item{Tm}{Team}
\item{Opp}{Opposition}
\item{GS}{Games started}
\item{MP}{Minutes played}
\item{FG}{Field goals}
\item{FGA}{Field goal attempts}
\item{FG\%}{Field goal percentage}
\item{3P}{3-point field goals}
\item{3PA}{3-point field goal attempts}
\item{3P\%}{Field goal percentage on 3-point field goal attempts}
\item{FT}{Free throws}
\item{FTA}{Free throw attempts}
\item{FT\%}{Free throw percentage}
\item{ORB}{Offensive rebounds}
\item{DRB}{Defensive rebounds}
\item{TRB}{Total rebounds}
\item{AST}{Assists}
\item{STL}{Steals}
\item{BLK}{Blocks}
\item{TOV}{Turnovers}
\item{PF}{Personal fouls}
\item{PTS}{Points}
\item{GmSc}{Game score}
\item{+/-}{Plus minus (possibly)}
\item{PlayerId}{https://www.basketball-reference.com/ player page identifier}
}
}
\description{
This function scrapes the per game player statistics for a single season.
}
