% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_season.R
\name{scrape_season_advanced}
\alias{scrape_season_advanced}
\title{Scrape player advanced statistics for a given season.}
\usage{
scrape_season_advanced(year)
}
\arguments{
\item{year}{The year that the season in question finished. i.e. 2015 would
return the statistics for the season 2014-15.}
}
\value{
A tibble with a row for each player.

The fields are:
\describe{
\item{Season}{The season these statistics are for}
\item{Player}{Player name}
\item{PlayerId}{https://www.basketball-reference.com/ player page identifier}
\item{Pos}{Position}
\item{Age}{Age of player at the start of February 1st of that season}
\item{Tm}{Team}
\item{G}{Games}
\item{PER}{Player efficiency rating: A measure of per-minute production
standardized such that the league average is 15.}
\item{TS\%}{True shooting percentage: A measure of shooting efficiency that
takes into account 2-point field goals, 3-point field goals, and free throws.}
\item{3PAr}{3-point attempt rate: The percentage of field goal attempts
taken from 3-point range}
\item{FTr}{Free throw attempt rate: The number of free throw attempts per field
goal attempt}
\item{ORB\%}{Offensive rebound percentage: An estimate of the percentage
of available offensive rebounds a player grabbed while he was on the floor}
\item{DRB\%}{Defensive rebound percentage: An estimate of the percentage
of available defensive rebounds a player grabbed while he was on the floor}
\item{TRB\%}{Total rebound percentage: An estimate of the percentage
of available rebounds a player grabbed while he was on the floor}
\item{AST\%}{Assist percentage: An estimate of the percentage of team-mate
field goals a player assisted while he was on the floor}
\item{STL\%}{Steal percentage: An estimate of the percentage of opponent
possessions that end with a steal by the player while he was on the floor}
\item{BLK\%}{Block percentage: An estimate of opponent two-point field goal
attempts blocked by the player while he was on the floor}
\item{TOV\%}{Turnover percentage: An estimate of turnovers committed per 100
plays}
\item{USG\%}{Usage percentage: An estimate of the percentage of team plays
used by a player while he was on the floor}
\item{OWS}{Offensive win shares: An estimate of the number of wins
contributed by a player due to his offense}
\item{DWS}{Defensive win shares: An estimate of the number of wins
contributed by a player due to his defense}
\item{WS}{Win shares: An estimate of the number of wins contributed by a
player}
\item{WS/48}{Win shares per 48 minutes: An estimate of the number of wins
contributed by a player per 48 minutes (the league average is .100)}
\item{OBPM}{Offensive box plus/minus: A box score estimate of the offensive
points per 100 possessions a player contributed above a league average
player, translated to an average team}
\item{DBPM}{Defensive box plus/minus: A box score estimate of the defensive
points per 100 possessions a player contributed above a league average
player, translated to an average team}
\item{BPM}{Box plus/minus: A box score estimate of the points per 100
possessions a player contributed above a league average player, translated to
an average team}
\item{VORP}{Value over replacement player: A box score estimate of the
points per 100 team possessions that a player contributed above a
replacement level (-2.0) player, translated to an average team and
prorated to an 82-game season. Multiply by 2.70 to convert to wins over
replacement.}
}
}
\description{
This function scrapes the advanced player statistics for a single season.
}
