% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_season.R
\name{scrape_season_per_100_poss}
\alias{scrape_season_per_100_poss}
\title{Scrape player per 100 possession statistics for a given season.}
\usage{
scrape_season_per_100_poss(year)
}
\arguments{
\item{year}{The year that the season in question finished. i.e. 2015 would
return the statistics for the season 2014-15.}
}
\value{
A tibble with a row for each player.

The fields are:
\describe{
\item{Season}{The season these statistics are for}
\item{Player}{Player name}
\item{PlayerId}{https://www.basketball-reference.com/ player page identifier}
\item{Pos}{Position}
\item{Age}{Age of player at the start of February 1st of that season}
\item{Tm}{Team}
\item{G}{Games}
\item{GS}{Games started}
\item{MP}{Minutes played}
\item{FG}{Field goals per 100 possessions}
\item{FGA}{Field goal attempts per 100 possessions}
\item{FG\%}{Field goal percentage}
\item{3P}{3-point field goals per 100 possessions}
\item{3PA}{3-point field goal attempts per 100 possessions}
\item{3P\%}{Field goal percentage on 3-point field goal attempts}
\item{2P}{2-point field goals per 100 possessions}
\item{2PA}{2-point field goal attempts per 100 possessions}
\item{2P\%}{Field goal percentage on 2-point field goal attempts}
\item{eFG\%}{Effective field goal percentage}
\item{FT}{Free throws per 100 possessions}
\item{FTA}{Free throw attempts per 100 possessions}
\item{FT\%}{Free throw percentage}
\item{ORB}{Offensive rebounds per 100 possessions}
\item{DRB}{Defensive rebounds per 100 possessions}
\item{TRB}{Total rebounds per 100 possessions}
\item{AST}{Assists per 100 possessions}
\item{STL}{Steals per 100 possessions}
\item{BLK}{Blocks per 100 possessions}
\item{TOV}{Turnovers per 100 possessions}
\item{PF}{Personal fouls per 100 possessions}
\item{PTS}{Points per 100 possessions}
\item{ORtg}{Offensive rating: An estimate of points produced per 100
possessions}
\item{DRtg}{Defensive rating: An estimate of points allowed per 100
possessions}
}
}
\description{
This function scrapes the per 100 possessions player statistics for a single
season.
}
