% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_season.R
\name{scrape_season_per_game}
\alias{scrape_season_per_game}
\title{Scrape player per game statistics for a given season.}
\usage{
scrape_season_per_game(year)
}
\arguments{
\item{year}{The year that the season in question finished. i.e. 2015 would
return the statistics for the season 2014-15.}
}
\value{
A tibble with a row for each player.

The fields are:
\describe{
\item{Season}{The season these statistics are for}
\item{Player}{Player name}
\item{PlayerId}{https://www.basketball-reference.com/ player page identifier}
\item{Pos}{Position}
\item{Age}{Age of player at the start of February 1st of that season}
\item{Tm}{Team}
\item{G}{Games}
\item{GS}{Games started}
\item{MP}{Minutes played}
\item{FG}{Field goals per game}
\item{FGA}{Field goal attempts per game}
\item{FG\%}{Field goal percentage}
\item{3P}{3-point field goals per game}
\item{3PA}{3-point field goal attempts per game}
\item{3P\%}{Field goal percentage on 3-point field goal attempts}
\item{2P}{2-point field goals per game}
\item{2PA}{2-point field goal attempts per game}
\item{2P\%}{Field goal percentage on 2-point field goal attempts}
\item{eFG\%}{Effective field goal percentage}
\item{FT}{Free throws per game}
\item{FTA}{Free throw attempts per game}
\item{FT\%}{Free throw percentage}
\item{ORB}{Offensive rebounds per game}
\item{DRB}{Defensive rebounds per game}
\item{TRB}{Total rebounds per game}
\item{AST}{Assists per game}
\item{STL}{Steals per game}
\item{BLK}{Blocks per game}
\item{TOV}{Turnovers per game}
\item{PF}{Personal fouls per game}
\item{PS/G}{Points per game}
}
}
\description{
This function scrapes the per game player statistics for a single season.
}
