% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_season.R
\name{scrape_season_totals}
\alias{scrape_season_totals}
\title{Scrape player total statistics for a given season.}
\usage{
scrape_season_totals(year)
}
\arguments{
\item{year}{The year that the season in question finished. i.e. 2015 would
return the statistics for the season 2014-15.}
}
\value{
A tibble with a row for each player.

The fields are:
\describe{
\item{Season}{The season these statistics are for}
\item{Player}{Player name}
\item{PlayerId}{https://www.basketball-reference.com/ player page identifier}
\item{Pos}{Position}
\item{Age}{Age of player at the start of February 1st of that season}
\item{Tm}{Team}
\item{G}{Games}
\item{GS}{Games started}
\item{MP}{Minutes played}
\item{FG}{Field goals}
\item{FGA}{Field goal attempts}
\item{FG\%}{Field goal percentage}
\item{3P}{3-point field goals}
\item{3PA}{3-point field goal attempts}
\item{3P\%}{Field goal percentage on 3-point field goal attempts}
\item{2P}{2-point field goals}
\item{2PA}{2-point field goal attempts}
\item{2P\%}{Field goal percentage on 2-point field goal attempts}
\item{eFG\%}{Effective field goal percentage}
\item{FT}{Free throws}
\item{FTA}{Free throw attempts}
\item{FT\%}{Free throw percentage}
\item{ORB}{Offensive rebounds}
\item{DRB}{Defensive rebounds}
\item{TRB}{Total rebounds}
\item{AST}{Assists}
\item{STL}{Steals}
\item{BLK}{Blocks}
\item{TOV}{Turnovers}
\item{PF}{Personal fouls}
\item{PTS}{Points}
}
}
\description{
This function scrapes the player statistics totals for a single season.
}
