% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_teams.R
\name{scrape_teams}
\alias{scrape_teams}
\title{Scrape franchise records.}
\usage{
scrape_teams(status = "active")
}
\arguments{
\item{status}{'active', 'defunct' or 'both'. 'active' returns only
those franchises that are still participating in the NBA. 'defunct' returns
only those franchises that no longer participate in the NBA. 'both' returns
all franchises.}
}
\value{
A tibble with a row for each team.

The fields are:

\describe{
\item{Franchise}{Franchise name}
\item{FranchiseId}{Three letter identifier}
\item{League}{The leagues the franchise participated in}
\item{From}{First year in league}
\item{To}{Last year in league}
\item{Years}{Years active}
\item{Games}{Total games played}
\item{Won}{Total wins}
\item{Lost}{Total losses}
\item{WinLossRatio}{Wins / Losses}
\item{PlayoffApps}{Total appearances in playoffs}
\item{DivisionWins}{Total divisional title wins}
\item{ConferenceWins}{Total conference title wins}
\item{ChampionshipWins}{Total championship title wins}
\item{Active}{True if franchise is still active in NBA.}
}
}
\description{
This function scrapes the NBA & ABA Team Index for overall records by
franchise.
}
