% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_players.R
\name{scrape_all_players}
\alias{scrape_all_players}
\title{Scrape all NBA players}
\usage{
scrape_all_players()
}
\value{
A tibble with a row for each player.

The fields are:

\describe{
\item{Player}{Player name}
\item{PlayerId}{Player page Id}
\item{From}{First year in league}
\item{To}{Last year in league}
\item{Position}{Position(s) played}
\item{Height}{Height in 'feet-inches' format}
\item{Weight}{Weight in pounds}
\item{DOB}{Date of birth}
\item{College}{What college did the player attend?}
\item{HOF}{Is this player in the Hall of Fame?}
}
}
\description{
This function scrapes the entire NBA & ABA player directory.
}
