% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_teams.R
\name{scrape_teams}
\alias{scrape_teams}
\title{Scrape franchise records.}
\usage{
scrape_teams(status = "active")
}
\arguments{
\item{status}{'active', 'defunct' or 'both'. 'active' returns only
those franchises that are still participating in the NBA. 'defunct' returns
only those franchises that no longer participate in the NBA. 'both' returns
all franchises.}
}
\value{
A tibble with a row for each team.

The fields are:

\describe{
\item{Franchise}{Franchise name}
\item{FranchiseId}{Three letter identifier}
\item{Lg}{The leagues the franchise participated in}
\item{From}{First year in league}
\item{To}{Last year in league}
\item{Yrs}{Years active}
\item{G}{Total games played}
\item{W}{Total wins}
\item{L}{Total losses}
\item{W/L\%}{Wins / Losses}
\item{Plyfs}{Total appearances in playoffs}
\item{Div}{Total divisional title wins (including ties)}
\item{Conf}{Total conference title wins}
\item{Champ}{Total championship title wins}
}
}
\description{
This function scrapes the NBA & ABA Team Index for overall records by
franchise.
}
